Moved taxonomy loading func to topic popup; Moved getFieldStyle() to matadataUpdate service.
