+++ angular/src/app/landing/keyword/keyword.component.html
@@ -1,12 +1,12 @@
 <div style="display: table">
         <div class="editable_container">
           <div class="editable_field" *ngIf="isEmpty; else hasKeyword">
-            <div class="editable_field" *ngIf="mdupdsvc.isEditMode" [ngStyle]="getFieldStyle()">
+            <div class="editable_field" *ngIf="mdupdsvc.isEditMode" [ngStyle]="mdupdsvc.getFieldStyle(fieldName)">
               Add subject keywords:&nbsp;&nbsp;
             </div>
           </div>
           <ng-template #hasKeyword>
-            <div class="editable_field" style="max-width:calc(100% - 4em);" [ngStyle]="getFieldStyle()">
+            <div class="editable_field" style="max-width:calc(100% - 4em);" [ngStyle]="mdupdsvc.getFieldStyle(fieldName)">
               <strong>Subject Keywords:</strong>
               <span class="font14 keywords" contenteditable="false"
                 *ngFor="let keyword of record['keyword']; let i =index">
